ok now i can get the test to fail

+++ src/grid_maker.h
@@ -266,7 +266,7 @@ class GridMaker {
      * @param[out] a 4D grid
      */
     template <typename Dtype>
-    CUDA_DEVICE_MEMBER void set_atoms(unsigned natoms, float3& grid_origin, 
+    CUDA_DEVICE_MEMBER void set_atoms(size_t natoms, float3& grid_origin, 
         const Grid<float, 2, true>& coords, const Grid<float, 1, true>& type_index, 
         const Grid<float, 1, true>& radii, Grid<Dtype, 4, true>& out);
 };
